Refactor: create useVerificationNoteFooter
