+++ src/components/VerificationNote/VerificationNoteFooter/index.tsx
@@ -1,87 +1,21 @@
 'use client';
 
-import { useState, useEffect } from 'react';
 import { FaLink } from 'react-icons/fa';
-
-import { useVerificationNoteStore } from '@/store/useVerificationNoteStore';
 import { ConfirmationModal } from '@/components/ConfirmationModal';
-import { notify } from '@/store/useToastStore';
-import { MODAL_MESSAGES, TOAST_MESSAGES } from '@/constants/Messages';
+import { MODAL_MESSAGES } from '@/constants/Messages';
+import { useVerificationNoteFooter } from '@/hooks/useVerificationNoteFooter';
 
 export const VerificationNoteFooter = () => {
   const {
     note,
-    imageUrl,
-    completePicName,
-    completeLink,
-    completeId,
-    setImageUrl,
-    setCompletePicName,
-    setNote,
-    setCompleteLink,
-    setCompleteId,
-  } = useVerificationNoteStore();
-
-  const [isSaved, setIsSaved] = useState(false);
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [confirmationMode, setConfirmationMode] = useState<'save' | 'load'>(
-    'save',
-  );
-
-  useEffect(() => {
-    const savedData = localStorage.getItem('verificationNote');
-    if (savedData) {
-      setIsSaved(true);
-    }
-  }, []);
-
-  const handleClick = () => {
-    if (isSaved) {
-      setConfirmationMode('load');
-    } else {
-      setConfirmationMode('save');
-    }
-    setIsModalOpen(true);
-  };
-
-  const handleCloseModal = () => {
-    setIsModalOpen(false);
-  };
-
-  const handleConfirm = () => {
-    if (confirmationMode === 'save') {
-      const dataToSave = {
-        imageUrl,
-        completePicName,
-        note,
-        completeLink,
-        completeId,
-      };
-      localStorage.setItem('verificationNote', JSON.stringify(dataToSave));
-
-      notify('success', TOAST_MESSAGES.SAVE_SUCCESS, 3000);
-      setIsSaved(true);
-    } else {
-      const savedData = localStorage.getItem('verificationNote');
-      if (savedData) {
-        const parsedData = JSON.parse(savedData);
-        setImageUrl(parsedData.imageUrl);
-        setCompletePicName(parsedData.completePicName);
-        setNote(parsedData.note);
-        setCompleteLink(parsedData.completeLink);
-        setCompleteId(parsedData.completeId);
-
-        localStorage.removeItem('verificationNote');
-        setIsSaved(false);
-        notify('success', TOAST_MESSAGES.LOAD_SUCCESS, 3000);
-      }
-    }
-    setIsModalOpen(false);
-  };
-
-  const handleCancel = () => {
-    setIsModalOpen(false);
-  };
+    isSaved,
+    isModalOpen,
+    confirmationMode,
+    handleClick,
+    handleCloseModal,
+    handleConfirm,
+    handleCancel,
+  } = useVerificationNoteFooter();
 
   return (
     <div className="mt-auto flex w-full items-center justify-between bg-custom-white-200 p-16">
